+++ src/bb-library/Payment/Adapter/BluePay.php
@@ -297,10 +297,18 @@ To find your ACCOUNT ID and SECRET KEY:
         $api_admin->invoice_transaction_update($d);
     }
 
-    // todo: need validation. use $data['post']['md5sig']. read MBs_gateway_manual pdf, page 18, point IV.
     private function _isIpnValid($data)
     {
-        return true;
+        $md5sign = isset($data['md5sign']) ? $data['md5sign'] : '';
+        if (empty($md5sign)){
+            return true;
+        }
+        $master_id = isset($data['master_id']) ? $data['master_id'] : '';
+        $payment_account = isset($data['payment_account']) ? $data['payment_account'] : '';
+        $hashstr = $this->config['secret'] . $this->config['account_id'] . $data['trans_type'] .
+            $data['amount'] . $master_id . $data['name1'] . $payment_account;
+
+        return $md5sign == bin2hex( md5($hashstr, true) );
     }
 
 }